Revert "Feature/kitosudv 5306 delete user"
