+++ Core.ApplicationServices/Users/Write/IUserWriteService.cs
@@ -16,5 +16,7 @@ namespace Core.ApplicationServices.Users.Write
 
         Result<UserCollectionPermissionsResult, OperationError> GetCollectionPermissions(
             Guid organizationUuid);
+
+        Maybe<OperationError> CopyUserRights(Guid organizationUuid, Guid fromUserUuid, Guid toUserUuid, UserRightsChangeParameters parameters);
     }
 }
