+++ Presentation.Web/Controllers/API/V2/Internal/Users/UsersInternalV2Controller.cs
@@ -2,6 +2,9 @@
 using Swashbuckle.Swagger.Annotations;
 using System.Net;
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Policy;
 using System.Web.Http;
 using Core.ApplicationServices.Users.Write;
 using Presentation.Web.Controllers.API.V2.Internal.Users.Mapping;
@@ -10,6 +13,11 @@ using Presentation.Web.Models.API.V2.Request.User;
 using System.Web.Http.Results;
 using Core.ApplicationServices;
 using Core.ApplicationServices.Model.Users;
+using Core.ApplicationServices.Model.Users.Write;
+using Core.ApplicationServices.Rights;
+using Core.DomainModel.Organization;
+using Core.DomainServices.Generic;
+using Presentation.Web.Models.API.V2.Internal.Request.User;
 
 namespace Presentation.Web.Controllers.API.V2.Internal.Users
 {
@@ -42,7 +50,7 @@ namespace Presentation.Web.Controllers.API.V2.Internal.Users
         [SwaggerResponse(HttpStatusCode.BadRequest)]
         [SwaggerResponse(HttpStatusCode.Unauthorized)]
         [SwaggerResponse(HttpStatusCode.Forbidden)]
-        public IHttpActionResult CreateUser([NonEmptyGuid] Guid organizationUuid, [FromBody] CreateUserRequestDTO parameters)
+        public IHttpActionResult CreateUnit([NonEmptyGuid] Guid organizationUuid, [FromBody] CreateUserRequestDTO parameters)
         {
             return _userWriteService.Create(organizationUuid, _writeModelMapper.FromPOST(parameters))
                 .Select(user => _userResponseModelMapper.ToUserResponseDTO(organizationUuid, user))
@@ -56,26 +64,12 @@ namespace Presentation.Web.Controllers.API.V2.Internal.Users
         [SwaggerResponse(HttpStatusCode.BadRequest)]
         [SwaggerResponse(HttpStatusCode.Unauthorized)]
         [SwaggerResponse(HttpStatusCode.Forbidden)]
-        public IHttpActionResult PatchUser([NonEmptyGuid] Guid organizationUuid, [NonEmptyGuid] Guid userUuid,
+        public IHttpActionResult PatchUnit([NonEmptyGuid] Guid organizationUuid, [NonEmptyGuid] Guid userUuid,
             [FromBody] UpdateUserRequestDTO parameters)
         {
             return _userWriteService.Update(organizationUuid, userUuid, _writeModelMapper.FromPATCH(parameters))
                 .Select(user => _userResponseModelMapper.ToUserResponseDTO(organizationUuid, user))
                 .Match(Ok, FromOperationError);
-        }
-
-        [Route("{userUuid}")]
-        [HttpDelete]
-        [SwaggerResponse(HttpStatusCode.OK)]
-        [SwaggerResponse(HttpStatusCode.NotFound)]
-        [SwaggerResponse(HttpStatusCode.BadRequest)]
-        [SwaggerResponse(HttpStatusCode.Unauthorized)]
-        [SwaggerResponse(HttpStatusCode.Forbidden)]
-        public IHttpActionResult DeleteUser([NonEmptyGuid] Guid organizationUuid, [NonEmptyGuid] Guid userUuid)
-        {
-            return _userService.DeleteUserByOrganizationUuid(userUuid, organizationUuid)
-                .Match(FromOperationError,
-                    Ok);
         }
 
         [Route("{userUuid}/notifications/send")]
@@ -117,6 +111,19 @@ namespace Presentation.Web.Controllers.API.V2.Internal.Users
                 .Match(Ok, FromOperationError);
         }
 
+        [Route("{fromUserUuid}/copy-roles/{toUserUuid}")]
+        [HttpPost]
+        [SwaggerResponse(HttpStatusCode.OK)]
+        [SwaggerResponse(HttpStatusCode.NotFound)]
+        [SwaggerResponse(HttpStatusCode.Unauthorized)]
+        [SwaggerResponse(HttpStatusCode.Forbidden)]
+        public IHttpActionResult CopyRoles([NonEmptyGuid] Guid organizationUuid, [NonEmptyGuid] Guid fromUserUuid, [NonEmptyGuid] Guid toUserUuid, [FromBody] CopyUserRightsRequestDTO request)
+        {
+            var parameters = MapCopyRightsDTOToParameters(request);
+            _userWriteService.CopyUserRights(organizationUuid, fromUserUuid, toUserUuid, parameters);
+            return Ok();
+        }
+
         private CreatedNegotiatedContentResult<UserResponseDTO> MapUserCreatedResponse(UserResponseDTO dto)
         {
             return Created($"{Request.RequestUri.AbsoluteUri.TrimEnd('/')}/{dto.Uuid}", dto);
@@ -127,6 +134,20 @@ namespace Presentation.Web.Controllers.API.V2.Internal.Users
         {
             return new UserCollectionPermissionsResponseDTO(permissions.Create, permissions.Edit, permissions.Delete);
 
+        }
+
+        private UserRightsChangeParameters MapCopyRightsDTOToParameters(CopyUserRightsRequestDTO request)
+        {
+            var unitRights = MapUserRightsDTOToRoleIdSet(request.UnitRights);
+            var systemRights = MapUserRightsDTOToRoleIdSet(request.SystemRights);
+            var contractRights = MapUserRightsDTOToRoleIdSet(request.ContractRights);
+            var dprRights = MapUserRightsDTOToRoleIdSet(request.DataProcessingRights);
+            return new UserRightsChangeParameters(new List<OrganizationRole>(), dprRights, systemRights, contractRights, unitRights);
+        }
+
+        private IEnumerable<int> MapUserRightsDTOToRoleIdSet(IEnumerable<CopyRightRequestDTO> rights)
+        {
+            return rights.Select(right => right.RoleId);
         }
     }
 }
