+++ Tests.Integration.Presentation.Web/Tests.Integration.Presentation.Web.csproj
@@ -160,7 +160,7 @@
     <Compile Include="Tools\Internal\Notifications\NotificationV2Helper.cs" />
     <Compile Include="Tools\Internal\Organizations\OrganizationInternalV2Helper.cs" />
     <Compile Include="Tools\Internal\Organizations\OrganizationUnitRegistrationV2Helper.cs" />
-    <Compile Include="Tools\Internal\Users\UsersInternalV2Helper.cs" />
+    <Compile Include="Tools\Internal\Users\UsersV2Helper.cs" />
     <Compile Include="Tools\OrganizationGridTestHelper.cs" />
     <Compile Include="Users\OrganizationRightApiTest.cs" />
     <Compile Include="Organizations\OrganizationUnitTests.cs" />
@@ -255,7 +255,7 @@
     <Compile Include="Users\UserPermissionsAPITest.cs" />
     <Compile Include="Users\UserRolesManagementApiTest.cs" />
     <Compile Include="Users\UserTest.cs" />
-    <Compile Include="Users\V2\UsersInternalApiV2Test.cs" />
+    <Compile Include="Users\V2\UserV2Test.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
