+++ Tests.Integration.Presentation.Web/Tools/Internal/Users/UsersV2Helper.cs
@@ -3,13 +3,14 @@ using System;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Presentation.Web.Models.API.V2.Internal.Request.User;
 using Presentation.Web.Models.API.V2.Request.User;
 using Xunit;
 using Presentation.Web.Models.API.V2.Internal.Response.User;
 
 namespace Tests.Integration.Presentation.Web.Tools.Internal.Users
 {
-    public static class UsersInternalV2Helper
+    public static class UsersV2Helper
     {
         public static async Task<UserResponseDTO> CreateUser(Guid organizationUuid, CreateUserRequestDTO request, Cookie cookie = null)
         {
@@ -21,19 +22,6 @@ namespace Tests.Integration.Presentation.Web.Tools.Internal.Users
             Assert.Equal(HttpStatusCode.Created, response.StatusCode);
 
             return await response.ReadResponseBodyAsAsync<UserResponseDTO>();
-        }
-
-        public static async Task<HttpStatusCode> DeleteUserAndVerifyStatusCode(Guid organizationUuid, Guid userUuid, Cookie cookie = null, HttpStatusCode expectedStatusCode = HttpStatusCode.OK)
-        {
-            var requestCookie = cookie ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl(
-                $"{ControllerPrefix(organizationUuid)}/{userUuid}");
-            using var response = await HttpApi.DeleteWithCookieAsync(url,
-                 requestCookie);
-            var statusCode = response.StatusCode;
-            Assert.Equal(expectedStatusCode, statusCode);
-
-            return statusCode;
         }
 
         public static async Task<UserResponseDTO> UpdateUser(Guid organizationUuid, Guid userUuid,
@@ -79,6 +67,14 @@ namespace Tests.Integration.Presentation.Web.Tools.Internal.Users
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
             return await response.ReadResponseBodyAsAsync<UserResponseDTO>();
+        }
+
+        public static async Task<HttpResponseMessage> CopyRoles(Guid organizationUuid, Guid fromUser, Guid toUser,
+            CopyUserRightsRequestDTO request)
+        {
+            var requestCookie = await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
+            var url = TestEnvironment.CreateUrl($"{ControllerPrefix(organizationUuid)}/{fromUser}/copy-roles/{toUser}");
+            return await HttpApi.PostWithCookieAsync(url, requestCookie, request);
         }
 
         private static string ControllerPrefix(Guid organizationUuid)
