+++ Tests.Integration.Presentation.Web/Users/V2/UserV2Test.cs
@@ -6,6 +6,9 @@ using Core.DomainModel.Organization;
 using Core.DomainModel;
 using Presentation.Web.Models.API.V1;
 using System.Threading.Tasks;
+using Presentation.Web.Controllers.API.V2.Internal.Mapping;
+using Presentation.Web.Models.API.V2.Internal.Request.User;
+using Presentation.Web.Models.API.V2.Internal.Response.Roles;
 using Presentation.Web.Models.API.V2.Internal.Response.User;
 using Presentation.Web.Models.API.V2.Request.User;
 using Tests.Integration.Presentation.Web.Tools;
@@ -15,7 +18,7 @@ using Xunit;
 
 namespace Tests.Integration.Presentation.Web.Users.V2
 {
-    public class UsersInternalApiV2Test : WithAutoFixture
+    public class UserV2Test : WithAutoFixture
     {
         [Fact]
         public async Task Can_Create_User()
@@ -25,7 +28,7 @@ namespace Tests.Integration.Presentation.Web.Users.V2
             var userRequest = CreateCreateUserRequest();
 
             //Act
-            var response = await UsersInternalV2Helper.CreateUser(organization.Uuid, userRequest);
+            var response = await UsersV2Helper.CreateUser(organization.Uuid, userRequest);
 
             //Assert
             AssertUserEqualsCreateRequest(userRequest, response);
@@ -36,11 +39,10 @@ namespace Tests.Integration.Presentation.Web.Users.V2
         {
             //Arrange
             var organization = await CreateOrganizationAsync();
-            var userRequest = CreateCreateUserRequest();
-            var user = await UsersInternalV2Helper.CreateUser(organization.Uuid, userRequest);
+            var user = await CreateUserAsync(organization.Uuid);
 
             //Act
-            var response = await UsersInternalV2Helper.SendNotification(organization.Uuid, user.Uuid);
+            var response = await UsersV2Helper.SendNotification(organization.Uuid, user.Uuid);
 
             //Assert
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
@@ -53,7 +55,7 @@ namespace Tests.Integration.Presentation.Web.Users.V2
             var organization = await CreateOrganizationAsync();
 
             //Act
-            var response = await UsersInternalV2Helper.GetUserCollectionPermissions(organization.Uuid);
+            var response = await UsersV2Helper.GetUserCollectionPermissions(organization.Uuid);
 
             //Assert
             Assert.True(response.Create);
@@ -66,11 +68,11 @@ namespace Tests.Integration.Presentation.Web.Users.V2
         {
             //Arrange
             var organization = await CreateOrganizationAsync();
-            var user = await UsersInternalV2Helper.CreateUser(organization.Uuid, CreateCreateUserRequest());
+            var user = await CreateUserAsync(organization.Uuid);
 
             //Act
             var updateRequest = A<UpdateUserRequestDTO>();
-            var response = await UsersInternalV2Helper.UpdateUser(organization.Uuid, user.Uuid, updateRequest);
+            var response = await UsersV2Helper.UpdateUser(organization.Uuid, user.Uuid, updateRequest);
 
             //Assert
             AssertUserEqualsUpdateRequest(updateRequest, response);
@@ -91,42 +93,29 @@ namespace Tests.Integration.Presentation.Web.Users.V2
             //Arrange
             var organization = await CreateOrganizationAsync();
             var organization2 = await CreateOrganizationAsync();
-            var userRequest = CreateCreateUserRequest();
-            var user = await UsersInternalV2Helper.CreateUser(organization.Uuid, userRequest);
+            var user = await CreateUserAsync(organization.Uuid);
 
             //Act
-            var response = await UsersInternalV2Helper.GetUserByEmail(organization2.Uuid, user.Email);
+            var response = await UsersV2Helper.GetUserByEmail(organization2.Uuid, user.Email);
 
             //Assert
             Assert.Equal(user.Uuid, response.Uuid);
         }
 
         [Fact]
-        public async Task Can_Delete_User()
+        public async Task Can_Copy_User_Roles()
         {
+            //Arrange
             var organization = await CreateOrganizationAsync();
-            var user = await UsersInternalV2Helper.CreateUser(organization.Uuid, CreateCreateUserRequest());
+            var fromUser = await CreateUserAsync(organization.Uuid);
+            var toUser = await CreateUserAsync(organization.Uuid);
+            var request = A<CopyUserRightsRequestDTO>();
 
-            _ = await UsersInternalV2Helper.DeleteUserAndVerifyStatusCode(organization.Uuid, user.Uuid);
-        }
+            //Act
+            var result = await UsersV2Helper.CopyRoles(organization.Uuid, fromUser.Uuid, toUser.Uuid, request);
 
-        [Fact]
-        public async Task Delete_User_Returns_Not_Found_If_Invalid_Org_Uuid()
-        {
-            var organization = await CreateOrganizationAsync();
-            var user = await UsersInternalV2Helper.CreateUser(organization.Uuid, CreateCreateUserRequest());
-            var invalidOrgUuid = new Guid();
-
-            _ = await UsersInternalV2Helper.DeleteUserAndVerifyStatusCode(invalidOrgUuid, user.Uuid, null, HttpStatusCode.NotFound);
-        }
-
-        [Fact]
-        public async Task Delete_User_Returns_Not_Found_If_Invalid_User_Uuid()
-        {
-            var organization = await CreateOrganizationAsync();
-            var invalidUserUuid = new Guid();
-
-            _ = await UsersInternalV2Helper.DeleteUserAndVerifyStatusCode(organization.Uuid, invalidUserUuid, null, HttpStatusCode.NotFound);
+            //Assert
+            Assert.Equal(HttpStatusCode.OK, result.StatusCode);
         }
 
         private void AssertUserEqualsUpdateRequest(UpdateUserRequestDTO request, UserResponseDTO response)
@@ -159,6 +148,12 @@ namespace Tests.Integration.Presentation.Web.Users.V2
             }
         }
 
+        private async Task<UserResponseDTO> CreateUserAsync(Guid organizationUuid)
+        {
+            return await UsersV2Helper.CreateUser(organizationUuid, CreateCreateUserRequest());
+        }
+        
+
         private async Task<OrganizationDTO> CreateOrganizationAsync()
         {
             var organizationName = CreateName();
@@ -169,7 +164,7 @@ namespace Tests.Integration.Presentation.Web.Users.V2
 
         private string CreateName()
         {
-            return $"{nameof(UsersInternalApiV2Test)}{A<string>()}";
+            return $"{nameof(UserV2Test)}{A<string>()}";
         }
 
         private string CreateEmail()
